requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
Flask-Cors==3.0.6
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
kaptan==0.5.10
libtmux==0.8.0
lxml==4.1.1
MarkupSafe==1.0
meld3==1.0.2
python-crontab==2.3.5
python-dateutil==2.7.3
PyYAML==3.13
requests==2.18.4
six==1.11.0
supervisor==3.3.4
tmuxp==1.4.1
urllib3==1.22
Werkzeug==0.12.2
